Added support for Profiling PS kernel (#7260)

+++ src/runtime_src/core/edge/ps_kernels/profiling/aie_trace_config/aie_trace_config.cpp
@@ -1,4 +1,5 @@
-/* Copyright (C) 2022 Advanced Micro Devices, Inc. - All rights reserved
+/**
+ * Copyright (C) 2022-2023 Advanced Micro Devices, Inc. - All rights reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may
  * not use this file except in compliance with the License. A copy of the
@@ -291,7 +292,7 @@ namespace {
       if ((numCoreCounters < config.coreCounterStartEvents.size())
           || (numMemoryCounters < config.memoryCounterStartEvents.size())) {
         
-        std::vector<uint32_t> src = {config.coreCounterStartEvents.size(), config.memoryCounterStartEvents.size(), col    , row + 1}; 
+        std::vector<uint32_t> src = {static_cast<uint32_t>(config.coreCounterStartEvents.size()), static_cast<uint32_t>(config.memoryCounterStartEvents.size()), col, static_cast<uint32_t>(row + 1)}; 
         addMessage(msgcfg, Messages::COUNTERS_NOT_RESERVED, src);
         releaseCurrentTileCounters(config);
         return 1;
@@ -340,7 +341,7 @@ namespace {
 
         auto ret = coreTrace->reserve();
         if (ret != XAIE_OK) {
-          std::vector<uint32_t> src = {col, row + 1, 0, 0};
+          std::vector<uint32_t> src = {col, static_cast<uint32_t>(row + 1), 0, 0};
           addMessage(msgcfg, Messages::CORE_MODULE_TRACE_NOT_RESERVED, src);
           releaseCurrentTileCounters(config);
           return 1;
@@ -368,7 +369,7 @@ namespace {
         coreEvents.clear();
         numTileCoreTraceEvents[numTraceEvents]++;
 
-        std::vector<uint32_t> src = {numTraceEvents, col, row, 0};
+        std::vector<uint32_t> src = {static_cast<uint32_t>(numTraceEvents), col, row, 0};
         addMessage(msgcfg, Messages::CORE_TRACE_EVENTS_RESERVED, src);
 
         if (coreTrace->setMode(XAIE_TRACE_EVENT_PC) != XAIE_OK) 
@@ -394,7 +395,7 @@ namespace {
 
         auto ret = memoryTrace->reserve();
         if (ret != XAIE_OK) {
-          std::vector<uint32_t> src = {col, row + 1, 0, 0};
+          std::vector<uint32_t> src = {col,  static_cast<uint32_t>(row + 1), 0, 0};
           addMessage(msgcfg, Messages::MEMORY_MODULE_TRACE_NOT_RESERVED, src);
           releaseCurrentTileCounters(config);
           return 1;
@@ -483,7 +484,7 @@ namespace {
         memoryEvents.clear();
         numTileMemoryTraceEvents[numTraceEvents]++;
 
-        std::vector<uint32_t> src = {numTraceEvents, col, row, 0};
+        std::vector<uint32_t> src = {static_cast<uint32_t>(numTraceEvents), col, row, 0};
         addMessage(msgcfg, Messages::MEMORY_TRACE_EVENTS_RESERVED, src);
 
         if (memoryTrace->setMode(XAIE_TRACE_EVENT_TIME) != XAIE_OK)
